Error HAndling for inputs

+++ app/javascript/components/SignUp.jsx
@@ -1,4 +1,4 @@
-import { Button, Card, Grid, Stack, TextField, Typography } from "@mui/material";
+import { Alert, Button, Card, Grid, Stack, TextField, Typography } from "@mui/material";
 import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { PRIMARY_COLOR_0, PRIMARY_COLOR_1 } from "./constants";
@@ -6,20 +6,30 @@ import { axiosInstance } from "./utils/axiosInstance";
 
 export default function SignUp(props) {
   const [user, setUser] = useState({});
+  const [error, setError] = useState({
+    status: false,
+    msg: "",
+  });
 
   const navigate = useNavigate();
 
   const handleSignUp = async () => {
     try {
-      let url = "/users";
-      let response = await axiosInstance.post(url, { user: user });
-      let data = await response.data;
-      if (data.success) {
-        localStorage.setItem("user", user.email);
-        props.setAuth(true);
-        navigate("/");
+      if (user.password === user.password_confirmation) {
+        let url = "/users";
+        let response = await axiosInstance.post(url, { user: user });
+        let data = await response.data;
+        if (data.success) {
+          localStorage.setItem("user", user.email);
+          props.setAuth(true);
+          navigate("/");
+        } else {
+          setError({ status: true, msg: data.error });
+        }
+        console.log(data);
+      } else {
+        setError({ status: true, msg: "Passowrds do not match!" });
       }
-      console.log(data);
     } catch (error) {
       console.log(error);
     }
@@ -30,28 +40,32 @@ export default function SignUp(props) {
   };
 
   return (
-    <Stack flexDirection={"row"} justifyContent={"center"} component={"nav"} padding={"20px"}>
+    <Stack flexDirection={"row"} justifyContent={"center"} padding={"20px"}>
       <Card sx={{ width: "300px", minHeight: "300px" }} raised>
         <Grid container spacing={3} direction={"column"} justifyContent={"center"} alignItems={"center"}>
+          {error.status && (
+            <Grid item sx={{ marginTop: "20px" }}>
+              <Alert severity='error'>{error.msg}</Alert>
+            </Grid>
+          )}
           <Grid item xs={12}>
             <TextField
               label='Email'
               type={"email"}
-              sx={{ marginTop: "30px" }}
-              onChange={(e) => {
-                setUser({ ...user, email: e.target.value });
-              }}
+              name='email'
+              sx={{ marginTop: "20px" }}
+              onChange={handleChange}
             ></TextField>
           </Grid>
           <Grid item xs={12}>
-            <TextField label='Password' name='password' type={"password"} onChange={(e) => handleChange(e)}></TextField>
+            <TextField label='Password' name='password' type={"password"} onChange={handleChange}></TextField>
           </Grid>
           <Grid item xs={12}>
             <TextField
               label='Confirm Password'
               name='password_confirmation'
               type={"password"}
-              onChange={(e) => handleChange(e)}
+              onChange={handleChange}
             ></TextField>
           </Grid>
           <Grid item xs={12}>
